use context to hold authentication data
