+++ pages/_app.js
@@ -7,6 +7,7 @@ import { useState, useEffect } from "react";
 import { useRouter } from "next/router";
 
 import supabase from "../utils/supabaseClient";
+import { AuthWrapper } from "../context/auth";
 
 function MyApp({ Component, pageProps }) {
   const [authenticatedState, setAuthenticatedState] =
@@ -21,7 +22,6 @@ function MyApp({ Component, pageProps }) {
         handleAuthChange(event, session);
         if (event === "SIGNED_IN") {
           setAuthenticatedState("authenticated");
-          // router.push("/profile-settings");
         }
         if (event === "SIGNED_OUT") {
           setAuthenticatedState("not-authenticated");
@@ -55,10 +55,12 @@ function MyApp({ Component, pageProps }) {
   }
 
   return (
-    <Layout>
-      {authenticatedState === "authenticated" && <Navbar></Navbar>}
-      <Component {...pageProps} />
-    </Layout>
+    <AuthWrapper>
+      <Layout>
+        {authenticatedState === "authenticated" && <Navbar></Navbar>}
+        <Component {...pageProps} />
+      </Layout>
+    </AuthWrapper>
   );
 }
