+++ pages/create-profile/index.js
@@ -3,51 +3,17 @@ import supabase from "../../utils/supabaseClient";
 import CurrentLocation from "../../components/CurrentLocation/CurrentLocation";
 import { useEffect, useState } from "react";
 import UserPhotoUpload from "../../components/UserProfile/UserPhotoUpload";
+import { useAuthContext } from "../../context/auth";
 
 // 1. To verify user's information (name, avatar_url, location, bio) is 100% filled.
 // 2. if(user information is 100%) then { redirect user to /groups}
 // 3. if(user infromation is missing something) then { return the edit form }
 
 function ProfileSettings(props) {
-  const [user, setUser] = useState(null);
-  const [loading, setLoading] = useState(true);
+  const { user } = useAuthContext();
   const [postcode, setPostcode] = useState(null);
   const [avatar, setAvatar] = useState(null);
   const [bio, setBio] = useState("");
-
-  useEffect(() => {
-    // Check active sessions and sets the user
-    const session = supabase.auth.session();
-
-    setUser(session?.user ?? null);
-    setLoading(false);
-
-    // Listen for changes on auth state (logged in, signed out, etc.)
-    const { data: listener } = supabase.auth.onAuthStateChange(
-      async (event, session) => {
-        setUser(session?.user ?? null);
-        setLoading(false);
-      }
-    );
-    return () => {
-      listener?.unsubscribe();
-    };
-  }, []);
-
-  //if the user just logged in this will be null, so we need to rely on the onAuthChanged,
-  //if the user has logged in before, supabase.auth.user() will have a user
-  useEffect(() => {
-    async function getUser() {
-      const user = supabase.auth.user();
-      //find in the db the matching profile
-      const { data, error } = await supabase
-        .from("profiles")
-        .select()
-        .eq("id", user.id);
-      setUser(data[0]);
-    }
-    getUser();
-  }, []);
 
   async function handleSubmit(event) {
     event.preventDefault();
